Added '_' and '|' operators, rearranged test classes.
